hells bells! WordPress does not support endpoints on custom taxonomies
There is a good way to do this;
register your taxonomy with 'ep_mask' = EP_CATEGORIES
but nobody does this.
it should be obligatory
it should be the default instead of EP_NONE
but because it is none of those, do a lot of magic to get custom taxonomies to work properly with an endpoint.

to make sure endpoints work, they are moved to the top.

Yes, I tried add_rewrite_rule(..., ..., 'top');
That doesn't work when already in the rewrite_rules_array filter.